BWAmap_BloodTranscriptomes
Because hemoglobin is so upregulated in the blood transcriptomes, we wanted to report the % of clean paired reads that map to hemoglobin. This script mapped clean paired reads to their corresponding transcriptome. The output files from this script should result in a counts table and stats file. *Performed on SuperComputer, Hopper (Auburn University)*.